+++ plugins/pretty/webpack.config.js
@@ -2,7 +2,7 @@ module.exports = [
   {
     target: 'web',
     entry: {
-      'pretty-plus/pretty-plus': `./plugins/pretty/src/pretty.ts`
+      'pretty/pretty': `./plugins/pretty/src/pretty.ts`
     }
   }
 ];