+++ CHANGELOG.md
@@ -9,3 +9,5 @@
 - [monaco]: minor tweak to server route
 - [monaco]: parse LESS with the LESS parser instead of CSS
 - [monaco-workers]: merge into main `monaco` plugin
+- [prettier]: `prettier` update (1.18.2 → 2.1.12)
+- [prettier]: adds XML plugin